requirements.txt
coqui-tts
soundfile
numpy
gradio
